+++ analysis/src/DocExtraction.ml
@@ -112,13 +112,13 @@ let rec stringifyDocItem ?(indentation = 0) ~originalEnv (item : docItem) =
       [
         ("id", Some (wrapInQuotes id));
         ("kind", Some (wrapInQuotes "value"));
-        ("name", Some (name |> Json.escape |> wrapInQuotes));
+        ("name", Some (name |> wrapInQuotes));
         ( "deprecated",
           match deprecated with
           | Some d -> Some (wrapInQuotes d)
           | None -> None );
         ( "signature",
-          Some (signature |> String.trim |> Json.escape |> wrapInQuotes) );
+          Some (signature |> String.trim |> wrapInQuotes) );
         ("docstrings", Some (stringifyDocstrings docstring));
       ]
   | Type {id; docstring; signature; name; deprecated; detail} ->
@@ -126,12 +126,12 @@ let rec stringifyDocItem ?(indentation = 0) ~originalEnv (item : docItem) =
       [
         ("id", Some (wrapInQuotes id));
         ("kind", Some (wrapInQuotes "type"));
-        ("name", Some (name |> Json.escape |> wrapInQuotes));
+        ("name", Some (name |> wrapInQuotes));
         ( "deprecated",
           match deprecated with
           | Some d -> Some (wrapInQuotes d)
           | None -> None );
-        ("signature", Some (signature |> Json.escape |> wrapInQuotes));
+        ("signature", Some (signature |> wrapInQuotes));
         ("docstrings", Some (stringifyDocstrings docstring));
         ( "detail",
           match detail with
@@ -294,17 +294,17 @@ let extractDocs ~path ~debug =
                    let id =
                      (modulePath |> List.rev |> List.hd) ^ "." ^ item.name
                    in
-                   let items =
+                   let items, internalDocstrings =
                      match
                        ProcessCmt.fileForModule ~package:full.package
                          aliasToModule
                      with
-                     | None -> []
+                     | None -> ([], [])
                      | Some file ->
                        let docs =
                          extractDocsForModule ~modulePath:[id] file.structure
                        in
-                       docs.items
+                       (docs.items, docs.docstring)
                    in
                    Some
                      (ModuleAlias
@@ -312,7 +312,7 @@ let extractDocs ~path ~debug =
                           id;
                           name = item.name;
                           items;
-                          docstring = item.docstring |> List.map String.trim;
+                          docstring = item.docstring @ internalDocstrings |> List.map String.trim;
                         })
                  | Module (Structure m) ->
                    (* module Whatever = {} in res or module Whatever: {} in resi. *)