+++ analysis/src/SharedTypes.ml
@@ -311,12 +311,19 @@ end = struct
     {env with exported = structure.exported; pathRev; parent = Some env}
 end
 
+type typeArgContext = {
+  env: QueryEnv.t;
+  typeArgs: Types.type_expr list;
+  typeParams: Types.type_expr list;
+}
+
 type polyVariantConstructor = {
   name: string;
   displayName: string;
   args: Types.type_expr list;
 }
 
+(* TODO(env-stuff) All envs for bool string etc can be removed. *)
 type innerType = TypeExpr of Types.type_expr | ExtractedType of completionType
 and completionType =
   | Tuple of QueryEnv.t * Types.type_expr list * Types.type_expr
@@ -331,13 +338,12 @@ and completionType =
   | Tbool of QueryEnv.t
   | Tarray of QueryEnv.t * innerType
   | Tstring of QueryEnv.t
+  | TtypeT of {env: QueryEnv.t; path: Path.t}
   | Tvariant of {
       env: QueryEnv.t;
       constructors: Constructor.t list;
       variantDecl: Types.type_declaration;
       variantName: string;
-      typeArgs: Types.type_expr list;
-      typeParams: Types.type_expr list;
     }
   | Tpolyvariant of {
       env: QueryEnv.t;
@@ -489,6 +495,7 @@ type builtInCompletionModules = {
 }
 
 type package = {
+  suffix: string;
   rootPath: filePath;
   projectFiles: FileSet.t;
   dependenciesFiles: FileSet.t;
@@ -553,7 +560,7 @@ let _ = locItemToString
 
 module Completable = struct
   (* Completion context *)
-  type completionContext = Type | Value | Module | Field
+  type completionContext = Type | Value | Module | Field | ValueOrField
 
   type argumentLabel =
     | Unlabelled of {argumentPosition: int}
@@ -619,8 +626,11 @@ module Completable = struct
 
   type patternMode = Default | Destructuring
 
+  type decoratorPayload = Module of string
+
   type t =
     | Cdecorator of string  (** e.g. @module *)
+    | CdecoratorPayload of decoratorPayload
     | CnamedArg of contextPath * string * string list
         (** e.g. (..., "label", ["l1", "l2"]) for ...(...~l1...~l2...~label...) *)
     | Cnone  (** e.g. don't complete inside strings *)
@@ -647,6 +657,7 @@ module Completable = struct
     | Type -> "Type"
     | Module -> "Module"
     | Field -> "Field"
+    | ValueOrField -> "ValueOrField"
 
   let rec contextPathToString = function
     | CPString -> "string"
@@ -701,6 +712,7 @@ module Completable = struct
   let toString = function
     | Cpath cp -> "Cpath " ^ contextPathToString cp
     | Cdecorator s -> "Cdecorator(" ^ str s ^ ")"
+    | CdecoratorPayload (Module s) -> "CdecoratorPayload(module=" ^ s ^ ")"
     | CnamedArg (cp, s, sl2) ->
       "CnamedArg("
       ^ (cp |> contextPathToString)
@@ -773,10 +785,11 @@ module Completion = struct
     docstring: string list;
     kind: kind;
     detail: string option;
+    typeArgContext: typeArgContext option;
   }
 
-  let create ~kind ~env ?(docstring = []) ?filterText ?detail ?deprecated
-      ?insertText name =
+  let create ~kind ~env ?typeArgContext ?(docstring = []) ?filterText ?detail
+      ?deprecated ?insertText name =
     {
       name;
       env;
@@ -788,10 +801,11 @@ module Completion = struct
       insertTextFormat = None;
       filterText;
       detail;
+      typeArgContext;
     }
 
-  let createWithSnippet ~name ?insertText ~kind ~env ?sortText ?deprecated
-      ?filterText ?detail ?(docstring = []) () =
+  let createWithSnippet ~name ?typeArgContext ?insertText ~kind ~env ?sortText
+      ?deprecated ?filterText ?detail ?(docstring = []) () =
     {
       name;
       env;
@@ -803,6 +817,7 @@ module Completion = struct
       insertTextFormat = Some Protocol.Snippet;
       filterText;
       detail;
+      typeArgContext;
     }
 
   (* https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_completion *)