+++ tools/CHANGELOG.md
@@ -12,6 +12,11 @@
 
 ## master
 
+#### :bug: Bug Fix
+
+- Fix docstrings for module alias. Get internal docstrings of module file. https://github.com/rescript-lang/rescript-vscode/pull/878
+- Fix extracted docs of types include escaped linebreaks in signature. https://github.com/rescript-lang/rescript-vscode/pull/891
+
 ## 0.3.0
 
 #### :rocket: New Feature